// lib/schemas/eventResponseDtoStatus.ts
/**
 * Generated by orval v8.1.0 🍺
 * Do not edit manually.
 * 5Sport API
 * API documentation for 5Sport authentication and services
 * OpenAPI spec version: 1.0
 */

/**
 * Trạng thái sự kiện
 */
export type EventResponseDtoStatus = typeof EventResponseDtoStatus[keyof typeof EventResponseDtoStatus];


export const EventResponseDtoStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  LIVE: 'LIVE',
  CLOSED: 'CLOSED',
  CANCELLED: 'CANCELLED',
} as const;